move integration test to solve circular dependency
