+++ Cargo.toml
@@ -41,3 +41,11 @@ parallel_hash = ["atelier-daemon/parallel_hash"]
 pretty_log = ["atelier-daemon/pretty_log"]
 rpc_io = ["atelier-loader/rpc_io"]
 handle = ["atelier-loader/handle"]
+
+[dev-dependencies]
+futures-core = "0.3.9"
+futures-io = "0.3.9"
+futures-util = "0.3.9"
+serde = "1.0.118"
+type-uuid = "0.1.2"
+uuid = "0.8.1"
