+++ daemon/Cargo.toml
@@ -52,7 +52,7 @@ pin-project = "0.4.8"
 uuid = { version = "0.8.1", features = ["v4"] }
 
 [dev-dependencies]
-tempfile = "3"
+tempfile = "3.1.0"
 
 [features]
 parallel_hash = ["rayon"]
