+++ loader/Cargo.toml
@@ -7,7 +7,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 crossbeam-channel = "0.5.0"
-
 atelier-core = { path = "../core", version = "0.1.0", features = ["serde-1"] }
 atelier-schema = { path = "../schema", version = "0.1.0", optional = true }
 tokio = { version = "0.2", features = ["tcp", "sync", "rt-core", "rt-util", "stream"], optional = true }
@@ -21,15 +20,6 @@ log = "0.4.11"
 dashmap = "4.0.1"
 serde = { version = "1.0.118", features = ["derive"], optional = true }
 uuid = { version = "0.8.1", optional = true }
-
-[dev-dependencies]
-atelier-daemon = { path = "../daemon", version = "0.1.0" }
-atelier-importer = { path = "../importer", version = "0.1.0" }
-uuid = { version = "0.8.1", features = ["v4"] }
-serde = { version = "1.0.118", features = ["derive"] }
-tokio = { version = "0.2", features = ["io-util"] }
-futures-io = { version = "0.3" }
-type-uuid = "0.1.2"
 
 [features]
 default = ["rpc_io", "handle"]
