--- loader/tests/assets/asset.txt.meta
@@ -1,29 +1,0 @@
-(
-    version: 1,
-    import_hash: Some(5294496694179974904),
-    importer_version: 1,
-    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
-    importer_options: (),
-    importer_state: (
-        id: Some("60352042-616f-460e-abd2-546195c060fe"),
-    ),
-    assets: [
-        (
-            id: "60352042-616f-460e-abd2-546195c060fe",
-            search_tags: [
-                ("file_name", Some("asset.txt")),
-            ],
-            build_pipeline: None,
-            artifact: Some((
-                hash: 10851886053989511955,
-                id: "60352042-616f-460e-abd2-546195c060fe",
-                build_deps: [],
-                load_deps: [],
-                compression: None,
-                compressed_size: Some(8),
-                uncompressed_size: Some(8),
-                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
-            )),
-        ),
-    ],
-)
