--- loader/tests/assets/asset_a.txt.meta
@@ -1,32 +1,0 @@
-(
-    version: 1,
-    import_hash: Some(13330951842867990526),
-    importer_version: 1,
-    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
-    importer_options: (),
-    importer_state: (
-        id: Some("a5ce4da0-675e-4460-be02-c8b145c2ee49"),
-    ),
-    assets: [
-        (
-            id: "a5ce4da0-675e-4460-be02-c8b145c2ee49",
-            search_tags: [
-                ("file_name", Some("asset_a.txt")),
-            ],
-            build_pipeline: None,
-            artifact: Some((
-                hash: 1356026748438998181,
-                id: "a5ce4da0-675e-4460-be02-c8b145c2ee49",
-                build_deps: [],
-                load_deps: [
-                    Uuid("039dc5f8-ee1c-4949-a7df-72383f12c7a2"),
-                    Uuid("55adb689-b91c-42a0-941b-de4a9f7f4f03"),
-                ],
-                compression: None,
-                compressed_size: Some(82),
-                uncompressed_size: Some(82),
-                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
-            )),
-        ),
-    ],
-)
