--- loader/tests/assets/asset_b.txt.meta
@@ -1,31 +1,0 @@
-(
-    version: 1,
-    import_hash: Some(11857256235859969024),
-    importer_version: 1,
-    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
-    importer_options: (),
-    importer_state: (
-        id: Some("039dc5f8-ee1c-4949-a7df-72383f12c7a2"),
-    ),
-    assets: [
-        (
-            id: "039dc5f8-ee1c-4949-a7df-72383f12c7a2",
-            search_tags: [
-                ("file_name", Some("asset_b.txt")),
-            ],
-            build_pipeline: None,
-            artifact: Some((
-                hash: 10402684315089908256,
-                id: "039dc5f8-ee1c-4949-a7df-72383f12c7a2",
-                build_deps: [],
-                load_deps: [
-                    Uuid("c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59"),
-                ],
-                compression: None,
-                compressed_size: Some(45),
-                uncompressed_size: Some(45),
-                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
-            )),
-        ),
-    ],
-)
