--- loader/tests/assets/asset_c.txt.meta
@@ -1,29 +1,0 @@
-(
-    version: 1,
-    import_hash: Some(16814324867712301873),
-    importer_version: 1,
-    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
-    importer_options: (),
-    importer_state: (
-        id: Some("c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59"),
-    ),
-    assets: [
-        (
-            id: "c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59",
-            search_tags: [
-                ("file_name", Some("asset_c.txt")),
-            ],
-            build_pipeline: None,
-            artifact: Some((
-                hash: 13583898190824372809,
-                id: "c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59",
-                build_deps: [],
-                load_deps: [],
-                compression: None,
-                compressed_size: Some(8),
-                uncompressed_size: Some(8),
-                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
-            )),
-        ),
-    ],
-)
