--- loader/tests/assets/asset_d.txt.meta
@@ -1,29 +1,0 @@
-(
-    version: 1,
-    import_hash: Some(11033065253965232501),
-    importer_version: 1,
-    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
-    importer_options: (),
-    importer_state: (
-        id: Some("55adb689-b91c-42a0-941b-de4a9f7f4f03"),
-    ),
-    assets: [
-        (
-            id: "55adb689-b91c-42a0-941b-de4a9f7f4f03",
-            search_tags: [
-                ("file_name", Some("asset_d.txt")),
-            ],
-            build_pipeline: None,
-            artifact: Some((
-                hash: 11215459537568283853,
-                id: "55adb689-b91c-42a0-941b-de4a9f7f4f03",
-                build_deps: [],
-                load_deps: [],
-                compression: None,
-                compressed_size: Some(8),
-                uncompressed_size: Some(8),
-                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
-            )),
-        ),
-    ],
-)
