+++ src/lib.rs
@@ -2,3 +2,318 @@ pub use atelier_core as core;
 pub use atelier_daemon as daemon;
 pub use atelier_importer as importer;
 pub use atelier_loader as loader;
+
+#[cfg(test)]
+mod tests {
+    use atelier_core::AssetUuid;
+    use atelier_core::{AssetRef, AssetTypeId};
+    use atelier_daemon::{init_logging, AssetDaemon};
+    use atelier_importer::{AsyncImporter, ImportedAsset, ImporterValue, Result as ImportResult};
+    use atelier_loader::{
+        storage::{
+            AssetLoadOp, AssetStorage, DefaultIndirectionResolver, LoadStatus, LoaderInfoProvider,
+        },
+        LoadHandle, Loader, RpcIO,
+    };
+    use futures_core::future::BoxFuture;
+    use futures_io::AsyncRead;
+    use futures_util::AsyncReadExt;
+    use serde::{Deserialize, Serialize};
+    use std::{
+        collections::HashMap,
+        iter::FromIterator,
+        path::PathBuf,
+        str::FromStr,
+        string::FromUtf8Error,
+        sync::RwLock,
+        thread::{self, JoinHandle},
+    };
+    use type_uuid::TypeUuid;
+    use uuid::Uuid;
+
+    #[derive(Debug)]
+    struct LoadState {
+        size: Option<usize>,
+        commit_version: Option<u32>,
+        load_version: Option<u32>,
+    }
+    struct Storage {
+        map: RwLock<HashMap<LoadHandle, LoadState>>,
+    }
+    impl AssetStorage for Storage {
+        fn update_asset(
+            &self,
+            _loader_info: &dyn LoaderInfoProvider,
+            _asset_type: &AssetTypeId,
+            data: Vec<u8>,
+            loader_handle: LoadHandle,
+            load_op: AssetLoadOp,
+            version: u32,
+        ) -> atelier_loader::Result<()> {
+            println!("update asset {:?} data size {}", loader_handle, data.len());
+            let mut map = self.map.write().unwrap();
+            let state = map.entry(loader_handle).or_insert(LoadState {
+                size: None,
+                commit_version: None,
+                load_version: None,
+            });
+
+            state.size = Some(data.len());
+            state.load_version = Some(version);
+            load_op.complete();
+            Ok(())
+        }
+        fn commit_asset_version(
+            &self,
+            _asset_type: &AssetTypeId,
+            loader_handle: LoadHandle,
+            version: u32,
+        ) {
+            println!("commit asset {:?}", loader_handle,);
+            let mut map = self.map.write().unwrap();
+            let state = map.get_mut(&loader_handle).unwrap();
+
+            assert!(state.load_version.unwrap() == version);
+            state.commit_version = Some(version);
+            state.load_version = None;
+        }
+        fn free(&self, _asset_type: &AssetTypeId, loader_handle: LoadHandle, _version: u32) {
+            println!("free asset {:?}", loader_handle);
+            self.map.write().unwrap().remove(&loader_handle);
+        }
+    }
+
+    /// Removes file comments (begin with `#`) and empty lines.
+    #[derive(Clone, Debug, Default, Deserialize, Serialize, TypeUuid)]
+    #[uuid = "346e6a3e-3278-4c53-b21c-99b4350662db"]
+    pub struct TxtFormat;
+    impl TxtFormat {
+        fn from_utf8(&self, vec: Vec<u8>) -> std::result::Result<String, FromUtf8Error> {
+            String::from_utf8(vec).map(|data| {
+                let processed = data
+                    .lines()
+                    .map(|line| {
+                        line.find('#')
+                            .map(|index| line.split_at(index).0)
+                            .unwrap_or(line)
+                            .trim()
+                    })
+                    .filter(|line| !line.is_empty())
+                    .flat_map(|line| line.chars().chain(std::iter::once('\n')));
+                String::from_iter(processed)
+            })
+        }
+    }
+    /// A simple state for Importer to retain the same UUID between imports
+    /// for all single-asset source files
+    #[derive(Default, Deserialize, Serialize, TypeUuid)]
+    #[uuid = "c50c36fe-8df0-48fe-b1d7-3e69ab00a997"]
+    pub struct TxtImporterState {
+        id: Option<AssetUuid>,
+    }
+    #[derive(TypeUuid)]
+    #[uuid = "fa50e08c-af6c-4ada-aed1-447c116d63bc"]
+    struct TxtImporter;
+    impl AsyncImporter for TxtImporter {
+        type State = TxtImporterState;
+        type Options = TxtFormat;
+
+        fn version_static() -> u32
+        where
+            Self: Sized,
+        {
+            1
+        }
+        fn version(&self) -> u32 {
+            Self::version_static()
+        }
+
+        fn import<'a>(
+            &'a self,
+            source: &'a mut (dyn AsyncRead + Unpin + Send + Sync),
+            txt_format: &'a Self::Options,
+            state: &'a mut Self::State,
+        ) -> BoxFuture<'a, ImportResult<ImporterValue>> {
+            Box::pin(async move {
+                if state.id.is_none() {
+                    state.id = Some(AssetUuid(*uuid::Uuid::new_v4().as_bytes()));
+                }
+                let mut bytes = Vec::new();
+                source.read_to_end(&mut bytes).await?;
+                let parsed_asset_data = txt_format
+                    .from_utf8(bytes)
+                    .expect("Failed to construct string asset.");
+
+                let load_deps = parsed_asset_data
+                    .lines()
+                    .filter_map(|line| Uuid::from_str(line).ok())
+                    .map(|uuid| AssetRef::Uuid(AssetUuid(*uuid.as_bytes())))
+                    .collect::<Vec<AssetRef>>();
+
+                Ok(ImporterValue {
+                    assets: vec![ImportedAsset {
+                        id: state.id.expect("AssetUuid not generated"),
+                        search_tags: Vec::new(),
+                        build_deps: Vec::new(),
+                        load_deps,
+                        asset_data: Box::new(parsed_asset_data),
+                        build_pipeline: None,
+                    }],
+                })
+            })
+        }
+    }
+
+    fn spawn_daemon(daemon_address: &str) -> JoinHandle<()> {
+        let daemon_address = daemon_address
+            .parse()
+            .expect("Failed to parse string as `SocketAddr`.");
+        thread::Builder::new()
+            .name("atelier-daemon".to_string())
+            .spawn(move || {
+                let tests_path = PathBuf::from_iter(&[env!("CARGO_MANIFEST_DIR"), "tests"]);
+
+                AssetDaemon::default()
+                    .with_db_path(tests_path.join("assets_db"))
+                    .with_address(daemon_address)
+                    .with_importer("txt", TxtImporter)
+                    .with_asset_dirs(vec![tests_path.join("assets")])
+                    .run();
+            })
+            .expect("Failed to spawn `atelier-daemon` thread.")
+    }
+
+    fn wait_for_status(
+        status: LoadStatus,
+        handle: LoadHandle,
+        loader: &mut Loader,
+        storage: &Storage,
+    ) {
+        loop {
+            println!(
+                "state {:?} expecting {:?}",
+                loader.get_load_status(handle),
+                status
+            );
+            if std::mem::discriminant(&status)
+                == std::mem::discriminant(&loader.get_load_status(handle))
+            {
+                break;
+            }
+            std::thread::sleep(std::time::Duration::from_millis(10));
+            if let Err(e) = loader.process(storage, &DefaultIndirectionResolver) {
+                println!("err {:?}", e);
+            }
+        }
+    }
+
+    #[ignore]
+    #[test]
+    fn test_connect() {
+        let _ = init_logging(); // Another test may have initialized logging, so we ignore errors.
+
+        // Start daemon in a separate thread
+        let daemon_port = 2500;
+        let daemon_address = format!("127.0.0.1:{}", daemon_port);
+        let _atelier_daemon = spawn_daemon(&daemon_address);
+
+        let mut loader = Loader::new(Box::new(RpcIO::new(daemon_address).unwrap()));
+        let handle = loader.add_ref(
+            // asset uuid of "tests/assets/asset.txt"
+            AssetUuid(
+                *uuid::Uuid::parse_str("60352042-616f-460e-abd2-546195c060fe")
+                    .unwrap()
+                    .as_bytes(),
+            ),
+        );
+        let storage = &mut Storage {
+            map: RwLock::new(HashMap::new()),
+        };
+        wait_for_status(LoadStatus::Loaded, handle, &mut loader, &storage);
+        loader.remove_ref(handle);
+        wait_for_status(LoadStatus::NotRequested, handle, &mut loader, &storage);
+    }
+
+    #[ignore]
+    #[test]
+    fn test_load_with_dependencies() {
+        let _ = init_logging(); // Another test may have initialized logging, so we ignore errors.
+
+        // Start daemon in a separate thread
+        let daemon_port = 2505;
+        let daemon_address = format!("127.0.0.1:{}", daemon_port);
+        let _atelier_daemon = spawn_daemon(&daemon_address);
+
+        let mut loader = Loader::new(Box::new(RpcIO::new(daemon_address).unwrap()));
+        let handle = loader.add_ref(
+            // asset uuid of "tests/assets/asset_a.txt"
+            AssetUuid(
+                *uuid::Uuid::parse_str("a5ce4da0-675e-4460-be02-c8b145c2ee49")
+                    .unwrap()
+                    .as_bytes(),
+            ),
+        );
+        let storage = &mut Storage {
+            map: RwLock::new(HashMap::new()),
+        };
+        wait_for_status(LoadStatus::Loaded, handle, &mut loader, &storage);
+
+        // Check that dependent assets are loaded
+        let asset_handles = asset_tree()
+            .iter()
+            .map(|(asset_uuid, file_name)| {
+                let asset_load_handle = loader
+                    .get_load(*asset_uuid)
+                    .unwrap_or_else(|| panic!("Expected `{}` to be loaded.", file_name));
+
+                (asset_load_handle, *file_name)
+            })
+            .collect::<Vec<(LoadHandle, &'static str)>>();
+
+        asset_handles
+            .iter()
+            .for_each(|(asset_load_handle, file_name)| {
+                assert_eq!(
+                    std::mem::discriminant(&LoadStatus::Loaded),
+                    std::mem::discriminant(&loader.get_load_status(*asset_load_handle)),
+                    "Expected `{}` to be loaded.",
+                    file_name
+                );
+            });
+
+        // Remove reference to top level asset.
+        loader.remove_ref(handle);
+        wait_for_status(LoadStatus::NotRequested, handle, &mut loader, &storage);
+
+        // Remove ref when unloading top level asset.
+        asset_handles
+            .iter()
+            .for_each(|(asset_load_handle, file_name)| {
+                println!("Waiting for {} to be `NotRequested`.", file_name);
+                wait_for_status(
+                    LoadStatus::NotRequested,
+                    *asset_load_handle,
+                    &mut loader,
+                    &storage,
+                );
+            });
+    }
+
+    fn asset_tree() -> Vec<(AssetUuid, &'static str)> {
+        [
+            ("a5ce4da0-675e-4460-be02-c8b145c2ee49", "asset_a.txt"),
+            ("039dc5f8-ee1c-4949-a7df-72383f12c7a2", "asset_b.txt"),
+            ("c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59", "asset_c.txt"),
+            ("55adb689-b91c-42a0-941b-de4a9f7f4f03", "asset_d.txt"),
+        ]
+        .iter()
+        .map(|(id, file_name)| {
+            let asset_uuid = *uuid::Uuid::parse_str(id)
+                .unwrap_or_else(|_| panic!("Failed to parse `{}` as `Uuid`.", id))
+                .as_bytes();
+
+            (AssetUuid(asset_uuid), *file_name)
+        })
+        .collect::<Vec<(AssetUuid, &'static str)>>()
+    }
+}
