+++ tests/assets/asset.txt.meta
@@ -1,0 +1,29 @@
+(
+    version: 1,
+    import_hash: Some(17364680304320274206),
+    importer_version: 1,
+    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
+    importer_options: (),
+    importer_state: (
+        id: Some("b24d209d-6622-4d78-a983-731e8b76f04d"),
+    ),
+    assets: [
+        (
+            id: "b24d209d-6622-4d78-a983-731e8b76f04d",
+            search_tags: [
+                ("file_name", Some("asset.txt")),
+            ],
+            build_pipeline: None,
+            artifact: Some((
+                id: (8678115335556946141),
+                asset_id: "b24d209d-6622-4d78-a983-731e8b76f04d",
+                build_deps: [],
+                load_deps: [],
+                compression: None,
+                compressed_size: Some(8),
+                uncompressed_size: Some(8),
+                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
+            )),
+        ),
+    ],
+)
