+++ tests/assets/asset_a.txt
