+++ tests/assets/asset_a.txt.meta
@@ -1,0 +1,32 @@
+(
+    version: 1,
+    import_hash: Some(16123910309833735425),
+    importer_version: 1,
+    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
+    importer_options: (),
+    importer_state: (
+        id: Some("d83bb247-2710-4c10-83df-d7daa53e19bf"),
+    ),
+    assets: [
+        (
+            id: "d83bb247-2710-4c10-83df-d7daa53e19bf",
+            search_tags: [
+                ("file_name", Some("asset_a.txt")),
+            ],
+            build_pipeline: None,
+            artifact: Some((
+                id: (10691359297162488684),
+                asset_id: "d83bb247-2710-4c10-83df-d7daa53e19bf",
+                build_deps: [],
+                load_deps: [
+                    Uuid("55adb689-b91c-42a0-941b-de4a9f7f4f03"),
+                    Uuid("039dc5f8-ee1c-4949-a7df-72383f12c7a2"),
+                ],
+                compression: None,
+                compressed_size: Some(82),
+                uncompressed_size: Some(82),
+                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
+            )),
+        ),
+    ],
+)
