+++ tests/assets/asset_b.txt
