+++ tests/assets/asset_b.txt.meta
@@ -1,0 +1,31 @@
+(
+    version: 1,
+    import_hash: Some(5444290597287671962),
+    importer_version: 1,
+    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
+    importer_options: (),
+    importer_state: (
+        id: Some("23da999a-a974-4d0e-918a-f226ea0b3e69"),
+    ),
+    assets: [
+        (
+            id: "23da999a-a974-4d0e-918a-f226ea0b3e69",
+            search_tags: [
+                ("file_name", Some("asset_b.txt")),
+            ],
+            build_pipeline: None,
+            artifact: Some((
+                id: (8172122769880646029),
+                asset_id: "23da999a-a974-4d0e-918a-f226ea0b3e69",
+                build_deps: [],
+                load_deps: [
+                    Uuid("c071f3ff-c9ea-4bf5-b3b9-bf5fc29f9b59"),
+                ],
+                compression: None,
+                compressed_size: Some(45),
+                uncompressed_size: Some(45),
+                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
+            )),
+        ),
+    ],
+)
