+++ tests/assets/asset_c.txt
