+++ tests/assets/asset_c.txt.meta
@@ -1,0 +1,29 @@
+(
+    version: 1,
+    import_hash: Some(11870937363006046355),
+    importer_version: 1,
+    importer_type: "fa50e08c-af6c-4ada-aed1-447c116d63bc",
+    importer_options: (),
+    importer_state: (
+        id: Some("40becaa7-cedb-466a-afee-41fecb1c916f"),
+    ),
+    assets: [
+        (
+            id: "40becaa7-cedb-466a-afee-41fecb1c916f",
+            search_tags: [
+                ("file_name", Some("asset_c.txt")),
+            ],
+            build_pipeline: None,
+            artifact: Some((
+                id: (14012283536715428919),
+                asset_id: "40becaa7-cedb-466a-afee-41fecb1c916f",
+                build_deps: [],
+                load_deps: [],
+                compression: None,
+                compressed_size: Some(8),
+                uncompressed_size: Some(8),
+                type_id: "7edbc10a-2147-499c-af9a-498723c7b35f",
+            )),
+        ),
+    ],
+)
