+++ tests/assets/asset_d.txt
